requirements.txt
tabula-py
JPype1
streamlit
pandas
transformers
torch
uvicorn
openai
fastapi